Make it easier to close a client/socket
Summary:
Currently closing a client is non-trivial: the `connect()` result must be cancelled if it is pending, but otherwise the resulting `socket` object must be closed.

This diff adds `RSocketClient#close()` that implements the above logic: cancel the underlying transport `connect()` if the transport is still waiting to connect, otherwise close the socket if it has been opened.

Reviewed By: mitermayer

Differential Revision: D5305382

fbshipit-source-id: 2c1df7e77effadccad9ee08e152980e1060c317d