Resumable DuplexConnection
Summary:
Implements `RSocketResumableTransport`, an implementation of the `DuplexConnection` interface that supports the RSocket resumption protocol.

This approach makes resumption largely transparent even to `RSocketClient`: the resumable transport is provided a source (function) for generating fresh underlying DuplexConnections. It maintains a buffer of pending/sent frames, watches the connection status of the underlying transport, and reconnects/resumes as necessary.

Reviewed By: yschimke

Differential Revision: D5343437

fbshipit-source-id: 30274479b0c73adfeeae5e090b3ee1d872ca8b0e