Support for RSocketWebSocketClient on nodeJS clients (#9)
Summary:
(following facebook internal discussion with lexs)

RSocketWebSocketClient currently relies on `WebSocket`, which is natively available in browsers. However in nodeJS, there is no native implementation of websockets, so they must be explicitly imported. Attempting to use this client as is in nodeJS throws a "WebSocket not defined" error.

I've updated the `RSocketWebSocketClient` constructor so that it now accepts a function, `wsCreator`, that takes `url` and `options` (similar to the WebSocket constructor), and returns a WebSocket object. This allows the user to pass any websocket constructor they want to RSocketWebSocketClient; for example, a nodeJS client can use [`ws`](https://www.npmjs.com/package/ws).

My change to the `RSocketWebSocketClient` constructor also allows one to pass headers to  `wsCreator`, through the `options` argument. This is a necessary change for some nodeJS clients to work correctly, though it is not currently possible, since the call to the `WebSocket` constructor only extracts `url` from `options`.

I don't have much experience with Flow, so any feedback would be great, particularly on typing things correctly. I'm opening this pull request to get the ball rolling.
Closes https://github.com/rsocket/rsocket-js/pull/9

Reviewed By: lexs

Differential Revision: D8680419

Pulled By: vsaraph

fbshipit-source-id: 09851fa4a4d2b27f69ac13baece7ff0506bb3466