Make LiteBuffer a proper class and typing (#20)
Summary:
Pull Request resolved: https://github.com/rsocket/rsocket-js/pull/20

LiteBuffer today still tries to be a Buffer polyfill, but fails to
provide some methods which makes it not a true polyfill. Instead, stop trying
and make it a separate class, that shared some methods with Buffer. We're
providing utilities to convert from and to Buffer/Uint8array depending on where
you're running.

Reviewed By: josephsavona

Differential Revision: D9456672

fbshipit-source-id: 18b4df85f6022aa0a09b6235fdc9b7ab604b4966